admin/requirements.txt
Django
djangorestframework==3.12.2
mysqlclient==2.2.4
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
python-dotenv